LayoutTests/tiled-drawing/scrolling/scroll-snap/scroll-snap-proximity-mainframe-expected.txt

Scrolling body with 2 drag ticks
- Did the scrolling snap to the top? YES
- Did scrolling snap to the second box? NO
Scrolling body with 30 drag ticks
- Did the scrolling snap to the top? NO
- Did scrolling snap to the second box? NO
Scrolling body with 60 drag ticks
- Did the scrolling snap to the top? NO
- Did scrolling snap to the second box? YES
